Backend/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
boto3==1.28.44
botocore==1.31.44
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coverage==7.3.1
cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.57.0
django-cors-headers==4.3.0
django-crontab==0.7.1
django-extensions==3.2.3
django-logging-middleware==0.1.2
django-ratelimit==4.1.0
django-rest-auth==0.9.5
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
graphviz==0.20.1
html5lib==1.1
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
loguru==0.7.2
lxml==4.9.3
multidict==6.0.4
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.2
PDFNetPython3==9.5.0.post1
Pillow==10.0.0
pluggy==1.3.0
protobuf==4.24.3
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==3.16.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
twilio==8.8.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
ua-parser==0.18.0
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.16
user-agents==2.2.0
webencodings==0.5.1
whitenoise==6.6.0
win32-setctime==1.1.0
xhtml2pdf==0.2.11
yarl==1.9.2
